Enables Continuous Integration with GitLab
This commit enables continuous integration (CI) when used with
[GitLab](https://gitlab.com/) (by adding a `.gitlab-ci.yml`
configuration file).

The definition says that it should automatically push to the npm
registry, if a new version is tagged.

It needs a **GitLab** Environment Variable defined here:
`Settings -> CI/CD -> Secret Variables`. The `key` should be `NPM_TOKEN` and it should
contain a valid **NPM_TOKEN** for pushing to the registry.

You can find one in your local `.npmrc`.
Just look for the `_authToken=[YOUR_API_KEY]` part.